Make waypoint navigation periodic
When port driven, the update hook only runs
when new trajectories or poses are received.
This works fine with periodic Vicon data or
when the rover is just constantly moving.
New motion commands are only written in the
update hook.

With our visual odometry, which only computes
a new pose while executing a motion command,
i.e., when it is moving, this does not work.

One of the two needs to be periodic.

Co-Author-By: Martin Azkarate <martin.azkarate@gmail.com>